Enable GPU gaming boost
This is specifically for our device as we only have 9 power levels (0-9).
This would probably be better implemented by reading the value of /sys/devices/fdb00000.qcom,kgsl-3d0/kgsl/kgsl-3d0/num_pwrlevels subtracting 1 as the values start at 0 but this sysfs entry only stores the number of values starting from 1.

The values would then need to be checked against in the activateGamingMode method instead of my hardcoded values.