Misc: Add Krait-C States.
This is a sloppy method to do this, but it works.

When I tried to do this via a for loop, it would only do the first 2 cores (even though the loop should have ran to 4).

I may revisit this in the future, but to the end user it'll be the same effect.