Downloader: Sloppy hack to fix downloads.
This just hardcodes the internal storage data path to fix auto download and flash.

This *REALLY* needs to be rewritten to properly get either the internal storage path via functions, or the appdata folder on the sdcard as we can no longer write to the sdcard where ever we want.