PerApp profiles: Automatic application of profiles based on foreground app
Introduction of PerApp profiles.

Notes:
This uses an accessibility service to get the foreground app. If it's not enabled, the app will present the page to the user.
Existing profiles need to be re-saved in order to be saved with a UUID
There is a slight delay while building the app list the first time (each open)