HBMWidget/AutoHBM: Allowing AutoHBM to resume after the second toggle of the widget
This should allow users to turn the HBM feature on or off, stopping AutoHBM temporarily. A previous commit reset the AutoHBM service after the screen was toggled.

This will allow the service to resume after the second toggle during that screen on cycle. I believe this would indicate that the user is ready for the service to resume