Utils: isLetter function
I've officially gotten tired of the Y/N 1/0 boolean toggles in sysfs paths.
This function can be used to pass a single character to check if it's a letter or digit.
That can in turn be used to check the existing sysfs value in an Iternary operation to select the right type of value.